Removed unwanted bug-file featurestests.feature.
Added {pause} in the end of given, when, then - lines in tests.py decorators and method parameter lists enabling execution to be potentially paused.
Added path fix to fileReader.py and fileexists check for tests.feature. This will cause the main to override the tests.feature with unrunnable code. the fileReader needs to be reworked.